services/pdf-generator/pyproject.toml
[project]
name = "invoicer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "pandas>=2.3.3",
    "pypdf>=6.5.0",
    "pyyaml>=6.0.3",
    "weasyprint>=67.0",
]
